Fix shortcuts and implement duplicate+copy/paste (#108)

+++ GloryEngine/Editor/GloryEditor/ObjectMenuCallbacks.cpp
@@ -15,6 +15,8 @@
 
 namespace Glory::Editor
 {
+	std::filesystem::path m_DeletingFile;
+
 	std::filesystem::path GetUnqiueFilePath(const std::filesystem::path& start)
 	{
 		std::filesystem::path currentPath = start;
@@ -36,15 +38,137 @@ namespace Glory::Editor
 	{
 		switch (currentMenu)
 		{
+		case T_SceneObject:
+		{
+			SceneObject* pSceneObject = (SceneObject*)pObject;
+			YAML::Emitter out;
+			out << YAML::BeginMap;
+			out << YAML::Key << "Type";
+			out << YAML::Value << "SceneObject";
+			out << YAML::Key << "Value";
+			out << YAML::Value << YAML::BeginSeq;
+			EditorSceneManager::SerializeObjects(pSceneObject, out);
+			out << YAML::EndSeq;
+			out << YAML::EndMap;
+			ImGui::SetClipboardText(out.c_str());
+			break;
+		}
+		case T_Resource:
+		{
+			const std::string& path = FileBrowserItem::GetHighlightedPath();
+			const UUID uuid = EditorAssetDatabase::FindAssetUUID(path);
 
+			YAML::Emitter out;
+			out << YAML::BeginMap;
+			out << YAML::Key << "Type";
+			out << YAML::Value << "Resource";
+			out << YAML::Key << "Value";
+			out << YAML::Value << YAML::BeginMap;
+			out << YAML::Key << "Path";
+			out << YAML::Value << path;
+			out << YAML::Key << "UUID";
+			out << YAML::Value << (uint64_t)uuid;
+			out << YAML::EndMap;
+			out << YAML::EndMap;
+			ImGui::SetClipboardText(out.c_str());
+
+			break;
+		}
+		case T_Folder:
+		{
+			const std::string& path = FileBrowserItem::GetHighlightedPath();
+			const UUID uuid = EditorAssetDatabase::FindAssetUUID(path);
+
+			YAML::Emitter out;
+			out << YAML::BeginMap;
+			out << YAML::Key << "Type";
+			out << YAML::Value << "Folder";
+			out << YAML::Key << "Value";
+			out << YAML::Value << YAML::BeginMap;
+			out << YAML::Key << "Path";
+			out << YAML::Value << path;
+			out << YAML::EndMap;
+			out << YAML::EndMap;
+			ImGui::SetClipboardText(out.c_str());
+			break;
+		}
 		}
 	}
 
 	OBJECTMENU_CALLBACK(PasteObjectCallback)
 	{
+		const char* clipboardText = ImGui::GetClipboardText();
+		YAML::Node clipboardNode;
+		try
+		{
+			clipboardNode = YAML::Load(clipboardText);
+		}
+		catch (const std::exception&)
+		{
+			Debug::LogError("Pasted object is not a YAML object!");
+			return;
+		}
+		if (!clipboardNode.IsDefined() || !clipboardNode.IsMap())
+		{
+			Debug::LogError("Pasted object is not a YAML object!");
+			return;
+		}
+
+		YAML::Node typeNode = clipboardNode["Type"];
+		YAML::Node valueNode = clipboardNode["Value"];
+
+		if (!typeNode.IsDefined() || !typeNode.IsScalar()) return;
+		const std::string& type = typeNode.as<std::string>();
+
 		switch (currentMenu)
 		{
+		case T_Scene:
+		{
+			if (type != "SceneObject") return;
+			GScene* pScene = (GScene*)pObject;
+			EditorSceneManager::PasteSceneObject(pScene, nullptr, valueNode);
+			break;
+		}
+		case T_Hierarchy:
+		case T_SceneObject:
+		{
+			if (type != "SceneObject") return;
+			GScene* pScene = EditorSceneManager::GetActiveScene();
+			if (!pScene) pScene = EditorSceneManager::NewScene();
+			SceneObject* pSceneObject = nullptr;
+			if (pObject)
+			{
+				pSceneObject = (SceneObject*)pObject;
+				pScene = pSceneObject->GetScene();
+			}
+			EditorSceneManager::PasteSceneObject(pScene, pSceneObject, valueNode);
+			break;
+		}
 
+		case T_Resource:
+		case T_Folder:
+		case T_ContentBrowser:
+		{
+			if (type != "Resource" && type != "Folder") return;
+			std::filesystem::path sourcePath = valueNode["Path"].as<std::string>();
+			std::filesystem::path destinationPath = FileBrowserItem::GetCurrentPath();
+			if (type == "Resource")
+			{
+				std::filesystem::path extension = sourcePath.extension();
+				destinationPath.append(sourcePath.filename().replace_extension().string() + " Copy").replace_extension(extension);
+				destinationPath = GetUnqiueFilePath(destinationPath);
+				std::filesystem::copy(sourcePath, destinationPath);
+				EditorAssetDatabase::ImportAssetAsync(destinationPath.string());
+			}
+			else if (type == "Folder")
+			{
+				destinationPath.append(sourcePath.filename().string() + " Copy");
+				destinationPath = GetUnqiueFilePath(destinationPath);
+				std::filesystem::copy(sourcePath, destinationPath);
+				EditorAssetDatabase::ImportAssetsAsync(destinationPath.string());
+			}
+			break;
+		}
 		}
 	}
 
@@ -52,7 +176,35 @@ namespace Glory::Editor
 	{
 		switch (currentMenu)
 		{
-
+		case T_SceneObject:
+		{
+			SceneObject* pSceneObject = (SceneObject*)pObject;
+			EditorSceneManager::DuplicateSceneObject(pSceneObject);
+			break;
+		}
+		case T_Resource:
+		{
+			std::filesystem::path path = FileBrowserItem::GetHighlightedPath();
+			std::filesystem::path copyPath = path;
+			std::filesystem::path extension = copyPath.extension();
+			copyPath.replace_filename(copyPath.filename().replace_extension().string() + " Copy").replace_extension(extension);
+			copyPath = GetUnqiueFilePath(copyPath);
+			std::filesystem::copy(path, copyPath);
+			EditorAssetDatabase::ImportAssetAsync(copyPath.string());
+			break;
+		}
+		case T_Folder:
+		{
+			std::filesystem::path path = FileBrowserItem::GetHighlightedPath();
+			std::filesystem::path copyPath = path;
+			copyPath.replace_filename(copyPath.filename().string() + " Copy");
+			copyPath = GetUnqiueFilePath(copyPath);
+			std::filesystem::copy(path, copyPath);
+			EditorAssetDatabase::ImportAssetsAsync(copyPath.string());
+			break;
+		}
+		default:
+			break;
 		}
 	}
 
@@ -74,33 +226,43 @@ namespace Glory::Editor
 
 		case ObjectMenuType::T_Resource:
 		{
+			m_DeletingFile = FileBrowserItem::GetHighlightedPath();
+			if (!std::filesystem::exists(m_DeletingFile))
+			{
+				m_DeletingFile = "";
+				return;
+			}
 			std::vector<std::string> buttons = {
 				"No",
 				"Yes",
 			};
 			std::vector<std::function<void()>> callbacks = {
-				[]() { PopupManager::CloseCurrentPopup(); },
-				[&]() { DeleteResource(); },
+				[]() { m_DeletingFile = ""; PopupManager::CloseCurrentPopup(); },
+				[&]() { DeleteResource(m_DeletingFile); m_DeletingFile = ""; PopupManager::CloseCurrentPopup(); },
 			};
 
-			std::filesystem::path file = FileBrowserItem::GetHighlightedPath();
-			std::string name = file.filename().replace_extension().string();
+			std::string name = m_DeletingFile.filename().replace_extension().string();
 			PopupManager::OpenModal("Delete " + name, "Are you sure you want to delete \"" + name + "\"?\nThis action cannot be undone!", buttons, callbacks);
 			break;
 		}
 
 		case ObjectMenuType::T_Folder:
 		{
-			std::filesystem::path path = FileBrowserItem::GetHighlightedPath();
+			m_DeletingFile = FileBrowserItem::GetHighlightedPath();
+			if (!std::filesystem::exists(m_DeletingFile))
+			{
+				m_DeletingFile = "";
+				return;
+			}
 			std::vector<std::string> buttons = {
 				"No",
 				"Yes",
 			};
 			std::vector<std::function<void()>> callbacks = {
-				[]() { PopupManager::CloseCurrentPopup(); },
-				[&]() { DeleteFolder(); },
+				[]() { m_DeletingFile = ""; PopupManager::CloseCurrentPopup(); },
+				[&]() { DeleteFolder(); m_DeletingFile = ""; PopupManager::CloseCurrentPopup(); },
 			};
-			PopupManager::OpenModal("Delete Folder", "Are you sure you want to delete the folder \"" + path.filename().string() + "\" and all assets underneath?\nThis action cannot be undone!", buttons, callbacks);
+			PopupManager::OpenModal("Delete Folder", "Are you sure you want to delete the folder \"" + m_DeletingFile.filename().string() + "\" and all assets underneath?\nThis action cannot be undone!", buttons, callbacks);
 			break;
 		}
 
@@ -275,8 +437,9 @@ namespace Glory::Editor
 
 	void DeleteFolder()
 	{
-		std::filesystem::path path = FileBrowserItem::GetHighlightedPath();
+		std::filesystem::path path = m_DeletingFile;
 		std::filesystem::path relativePath = path.lexically_relative(Game::GetAssetPath());
+		if (!std::filesystem::exists(path)) return;
 		if (!std::filesystem::remove_all(path)) return;
 
 		EditorAssetDatabase::DeleteAssets(relativePath.string());
@@ -285,15 +448,14 @@ namespace Glory::Editor
 		PopupManager::CloseCurrentPopup();
 	}
 
-	void DeleteResource()
+	void DeleteResource(std::filesystem::path path)
 	{
-		std::filesystem::path path = FileBrowserItem::GetHighlightedPath();
+		if (path.empty()) path = FileBrowserItem::GetHighlightedPath();
 		std::filesystem::path relativePath = path.lexically_relative(Game::GetAssetPath());
 		if (!std::filesystem::remove(path)) return;
 
 		EditorAssetDatabase::DeleteAsset(relativePath.string());
 		FileBrowserItem::GetSelectedFolder()->Refresh();
 		FileBrowserItem::GetSelectedFolder()->SortChildren();
-		PopupManager::CloseCurrentPopup();
 	}
 }
